Change how provider options are set
In another CAIDA codebase, there's the following comment:

/* One day, someone might update libipmeta to not take horrible getopt
 * style configuration and then these nasty functions could go away. */

This patch provides this change so that instead of providing providers
with getopt style options, we use a varargs method of handling options.

This change makes it so that the init functions for providers accept new
options, the optstr, and a variable amount of strings that hold the
options. The optstr is what the getopt flags were (-d, -l, etc), except
as a string of these options.

An example of how this is used as follows:

ipmeta_provider_providername_init(provider, "lb", "/path/to/locfile",
                                  "/path/to/blkfile");